+++ include/LabSound/core/AudioIOCallback.h
@@ -15,7 +15,7 @@ struct AudioIOCallback
 {
     // render() is called periodically to get the next render quantum of audio into destinationBus.
     // Optional audio input is given in sourceBus (if it's not 0).
-    virtual void render(AudioBus * sourceBus, AudioBus * destinationBus, size_t framesToProcess) = 0;
+    virtual void render(AudioBus * sourceBus, AudioBus * destinationBus, uint32_t framesToProcess) = 0;
     virtual ~AudioIOCallback() {}
 };
